+++ crates/koharu-runtime/src/runtime/packages/diffusion.rs
@@ -12,7 +12,7 @@ use crate::{
     source::extract,
 };
 
-const RELEASE: &str = "master-849-d04e895.1";
+const RELEASE: &str = "master-853-b68d586";
 
 #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq, strum::Display, strum::EnumProperty)]
 pub(crate) enum Diffusion {