+++ crates/koharu-runtime/src/runtime/packages/llama.rs
@@ -12,7 +12,7 @@ use crate::{
     source::extract,
 };
 
-const RELEASE: &str = "b10872";
+const RELEASE: &str = "b10903";
 
 #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq, strum::Display, strum::EnumProperty)]
 pub(crate) enum Llama {